refactor: create TwitchHelixRateLimitTracker
